Fourth Way, using Function.prototype.bind()
Create the function to be called outside the loop. While sending it as a callback, bind it to `null` with an initial argument as  `i`. 

Why does it work? Basically, we are not sending the original function as `setTimeout`'s callback but actually sending the return value from the bind operation, and that return value is the copy of the function. Also, bind allows us to not immediately call the function like .call() or .apply() does. 

Code: 

```js

function display(x){
  console.log(x)
}
for (var i = 0; i < 4; i++) {
  setTimeout(display.bind(null,i), 0)
}
```